pyproject.toml
[tool.poetry]
name = "sparkflow"
version = "0.1.0"
description = ""
authors = ["Vugar Abdulali <vabdulaliyev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
pyspark = "3.5.0"
pyyaml = "^6.0"
pydantic = "^2.10.3"
jinja2 = "^3.1.5"
matplotlib = "^3.10.1"
networkx = "^3.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
